move coursegroups functions to own controller
